[update] Move utils to package.
