+++ salesforce/utils/__init__.py
@@ -1,0 +1,6 @@
+# encoding: utf-8
+from .decorators import *
+from .reqs import *
+from .rest import *
+from .soap import *
+from .validators import *
